refactor LottoController -> LottoService, LottoMain -> LottoController 이름 변경
