+++ src/main/java/LottoController.java
@@ -1,33 +1,20 @@
 import java.util.List;
+import java.util.Set;
 
-import domain.LottoGame;
-import domain.LottoResults;
 import domain.ManualRequest;
-import domain.WinningStatistics;
 import view.LottoInputView;
-import view.LottoOutputView;
 
 public class LottoController {
-    private LottoGame lottoGame;
 
-    public void playLottoGames(ManualRequest manualRequest, int money) {
-        lottoGame = new LottoGame();
-        lottoGame.generateLottoResultsFromMoney(manualRequest, money);
-    }
-
-    public void lottoResults() {
-        LottoResults lottoResults = lottoGame.getLottoResults();
-        int gameCount = lottoGame.getCount();
-        LottoOutputView.printLottoResults(gameCount, lottoResults);
-    }
-
-    public void winningStatistics() {
-        WinningStatistics winningStatistics = lottoGame.calculateWinningStatistics(LottoInputView.getWinningNumbers(), LottoInputView.getBonusNumber());
-        LottoOutputView.printBeforeWinnings();
-        LottoOutputView.printWinnings(winningStatistics);
-    }
-
-    public void returnOnInvestment() {
-        LottoOutputView.printReturnOnInvestment(lottoGame.getReturnOnInvestment());
+    public static void main(String[] args) {
+        int money = LottoInputView.getMoney();
+        int manualCount = LottoInputView.getManualCount();
+        List<Set<Integer>> manualNumbers = LottoInputView.getManualLottos(manualCount);
+        LottoService lottoService = new LottoService();
+        ManualRequest manualRequest = new ManualRequest(manualCount, manualNumbers);
+        lottoService.playLottoGames(manualRequest, money);
+        lottoService.lottoResults();
+        lottoService.winningStatistics();
+        lottoService.returnOnInvestment();
     }
 }
